+++ src/onepay/PaymentGateway.php
@@ -177,9 +177,11 @@ class PaymentGateway extends BasePaymentGateway
     protected function getVerifyRequestData($command, \yii\web\Request $request): array
     {
         $params = [
-            'vpc_Command', 'vpc_Locale', 'vpc_MerchTxnRef', 'vpc_Merchant', 'vpc_OrderInfo', 'vpc_Amount',
-            'vpc_TxnResponseCode', 'vpc_TransactionNo', 'vcp_Message', 'vpc_SecureHash', 'vpc_AcqResponseCode',
-            'vpc_Authorizeld', 'vpc_Card', 'vpc_3DSECI', 'vpc_3Dsenrolled', 'vpc_3Dsstatus', 'vpc_CommercialCard'
+            'vpc_Command', 'vpc_Locale', 'vpc_MerchTxnRef', 'vpc_TxnResponseCode',
+            'vpc_OrderInfo', 'vpc_Amount', 'vpc_SecureHash', 'vpc_AcqResponseCode',
+            'vpc_TransactionNo', 'vcp_Message', 'vpc_3Dsstatus', 'vpc_3Dsstatus',
+            'vpc_Authorizeld', 'vpc_Card', 'vpc_3DSECI', 'vpc_Merchant', 'vpc_Version',
+            'vpc_CurrencyCode', 'vpc_AdditionData', 'vpc_CommercialCard', 'vpc_3Dsenrolled'
         ];
 
         $data = [];